Fix data download workflow.
Attempt to fix data download workflow. Particularly errors caused by
use of "Pkg" functions. These may be caused by the Julia version being
used, which was 1.1. It appears that despite the version specification
being set to 1.10, if the characters were not escaped with "1.10" the
trailing 0 was being removed an an improper version being set.

Hopefully by escaping the full version string it will fix that.